% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_clock_vs_calendar}
\alias{plot_clock_vs_calendar}
\title{Event-clock time versus calendar time}
\usage{
plot_clock_vs_calendar(path, ...)
}
\arguments{
\item{path}{An \code{event_clock_path} object from \code{\link[=event_clock_path]{event_clock_path()}} (an
\code{event_prices} object is converted automatically).}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plots the share of clock time elapsed against the share of calendar
time elapsed. The 45-degree line is the "information arrives uniformly"
benchmark; a curve below it means information is back-loaded (it waits
for the deadline), above it front-loaded.
}
\examples{
data(us2016)
ep <- as_event_prices(us2016, time = "date", price = "trump",
                      event_date = as.Date("2016-11-08"))
plot_clock_vs_calendar(event_clock_path(ep))
}
