% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{plot_signature}
\alias{plot_signature}
\title{Plot the sampling-frequency signature}
\usage{
plot_signature(sig, ...)
}
\arguments{
\item{sig}{An \code{ec_signature} object from \code{\link[=ec_signature]{ec_signature()}} (an
\code{event_prices} object is converted automatically).}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot the sampling-frequency signature
}
\examples{
data(polymarket2024)
plot_signature(ec_signature(as_event_prices(polymarket2024), max_every = 24))
}
