% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_q}
\alias{plot_q}
\title{Plot the event-probability path}
\usage{
plot_q(x, event_date = NULL, ...)
}
\arguments{
\item{x}{An \code{event_prices} object (or coercible).}

\item{event_date}{Optional event date; defaults to the one stored on
\code{x}.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Line plot of the traded event probability \eqn{q_t}, with an optional
marker at the scheduled event date.
}
\examples{
data(brexit2016)
ep <- as_event_prices(brexit2016, time = "date", price = "q_leave",
                      event_date = as.Date("2016-06-23"))
plot_q(ep)
}
